Update PHP Code
For Use Case 1 & 3

+++ assets/php/removeMaterial.php
@@ -1,7 +1,4 @@
 <?php
-//declare error count for checking the existing of user/email
-$errors = array();
-
 //get the data from manageMaterial.php
 $materialID = $_POST["materialID"];
 
@@ -10,14 +7,22 @@ if ($conn->connect_error){
 	die("Connection failure");
 }
 
-//use Material table
-$materialTable = "use material";
-$result = $conn->query($materialTable);
-
-//delete the material 
-$deleteData = "DELETE FROM material where MATERIAL_ID='$materialID';";
-if ($conn->query($deleteData)==TRUE){
+//check if the selected material can be removed or not
+//since some collectors may hold it in their material collection
+$sql = "SELECT * FROM collectormaterial WHERE MATERIAL_ID='$materialID'";
+$result = mysqli_query($conn, $sql);
+if(mysqli_num_rows($result) != 0)
+{
 	echo "<script type='text/javascript'>
-	alert('Selected material has been removed successfully!');
-	window.location = '/BIT216/manageMaterial.php'; </script>";}
+	alert('Sorry, material cannot be removed because some collectors have holded it.');
+	window.location = '/BIT216/manageMaterial.php'; </script>";
+}
+else{
+//delete the material
+	$deleteData = "DELETE FROM material where MATERIAL_ID='$materialID';";
+	if ($conn->query($deleteData)==TRUE){
+		echo "<script type='text/javascript'>
+		alert('Selected material has been removed successfully!');
+		window.location = '/BIT216/manageMaterial.php'; </script>";}
+}
 ?>
